Renamed heap_manager files
