+++ src/dataclay/executionenv/server/ExecutionEnvironmentSrv.py
@@ -31,7 +31,7 @@ from dataclay.runtime import settings
 from dataclay.communication.grpc.clients.LogicModuleGrpcClient import LMClient
 from dataclay.communication.grpc.clients.StorageLocationGrpcClient import SLClient
 from dataclay.communication.grpc.server.execution_environment_servicer import DataServiceEE
-from dataclay.executionenv.ExecutionEnvironment import ExecutionEnvironment
+from dataclay.executionenv.execution_environment import ExecutionEnvironment
 from dataclay.util import Configuration
 from dataclay.util.classloaders import (
     ClassLoader,
