--- src/dataclay/managers/classes.py
@@ -1,15 +1,0 @@
-""" Class description goes here. """
-
-"""
-Created on 2 ene. 2018
-
-@author: daniel
-"""
-
-"""
-
-WARNING: THIS CLASS IS MANTAINED ONLY BECAUSE MF2C DEMO IS IMPORTING MODULES FROM HERE AND IT IS WRONG!
-
-"""
-
-from dataclay.dataclay_object import DataClayObject
